executor: Remove wasmi-specific code from wasm_runtime.
