+++ core/executor/src/wasmi_execution.rs
@@ -19,18 +19,21 @@
 //! This module defines and implements the wasm part of Substrate Host Interface and provides
 //! an interface for calling into the wasm runtime.
 
-use std::str;
+use std::{str, mem};
 use wasmi::{
 	Module, ModuleInstance, MemoryInstance, MemoryRef, TableRef, ImportsBuilder, ModuleRef,
 	memory_units::Pages, RuntimeValue::{I32, I64, self},
 };
-use crate::error::{Error, Result};
+use crate::error::{Error, WasmError};
 use codec::{Encode, Decode};
 use primitives::{sandbox as sandbox_primitives, Blake2Hasher, traits::Externalities};
 use crate::host_interface::SubstrateExternals;
 use crate::sandbox;
 use crate::allocator;
+use crate::wasm_runtime::WasmRuntime;
 use log::trace;
+use parity_wasm::elements::{deserialize_buffer, DataSegment, Instruction, Module as RawModule};
+use runtime_version::RuntimeVersion;
 use wasm_interface::{
 	FunctionContext, HostFunctions, Pointer, WordSize, Sandbox, MemoryId, Result as WResult,
 };
@@ -43,7 +46,7 @@ struct FunctionExecutor {
 }
 
 impl FunctionExecutor {
-	fn new(m: MemoryRef, heap_base: u32, t: Option<TableRef>) -> Result<Self> {
+	fn new(m: MemoryRef, heap_base: u32, t: Option<TableRef>) -> Result<Self, Error> {
 		Ok(FunctionExecutor {
 			sandbox_store: sandbox::Store::new(),
 			heap: allocator::FreeingBumpHeapAllocator::new(heap_base),
@@ -62,22 +65,22 @@ impl sandbox::SandboxCapabilities for FunctionExecutor {
 	fn store_mut(&mut self) -> &mut sandbox::Store<Self::SupervisorFuncRef> {
 		&mut self.sandbox_store
 	}
-	fn allocate(&mut self, len: WordSize) -> Result<Pointer<u8>> {
+	fn allocate(&mut self, len: WordSize) -> Result<Pointer<u8>, Error> {
 		let heap = &mut self.heap;
 		self.memory.with_direct_access_mut(|mem| {
 			heap.allocate(mem, len)
 		})
 	}
-	fn deallocate(&mut self, ptr: Pointer<u8>) -> Result<()> {
+	fn deallocate(&mut self, ptr: Pointer<u8>) -> Result<(), Error> {
 		let heap = &mut self.heap;
 		self.memory.with_direct_access_mut(|mem| {
 			heap.deallocate(mem, ptr)
 		})
 	}
-	fn write_memory(&mut self, ptr: Pointer<u8>, data: &[u8]) -> Result<()> {
+	fn write_memory(&mut self, ptr: Pointer<u8>, data: &[u8]) -> Result<(), Error> {
 		self.memory.set(ptr.into(), data).map_err(Into::into)
 	}
-	fn read_memory(&self, ptr: Pointer<u8>, len: WordSize) -> Result<Vec<u8>> {
+	fn read_memory(&self, ptr: Pointer<u8>, len: WordSize) -> Result<Vec<u8>, Error> {
 		self.memory.get(ptr.into(), len as usize).map_err(Into::into)
 	}
 
@@ -88,7 +91,7 @@ impl sandbox::SandboxCapabilities for FunctionExecutor {
 		invoke_args_len: WordSize,
 		state: u32,
 		func_idx: sandbox::SupervisorFuncIndex,
-	) -> Result<i64>
+	) -> Result<i64, Error>
 	{
 		let result = wasmi::FuncInstance::invoke(
 			dispatch_thunk,
@@ -299,7 +302,7 @@ impl FunctionExecutor {
 
 impl wasmi::Externals for FunctionExecutor {
 	fn invoke_index(&mut self, index: usize, args: wasmi::RuntimeArgs)
-		-> std::result::Result<Option<wasmi::RuntimeValue>, wasmi::Trap>
+		-> Result<Option<wasmi::RuntimeValue>, wasmi::Trap>
 	{
 		let mut args = args.as_ref().iter().copied().map(Into::into);
 		let function = SubstrateExternals::functions().get(index).ok_or_else(||
@@ -327,14 +330,14 @@ pub fn call(
 	code: &[u8],
 	method: &str,
 	data: &[u8],
-) -> Result<Vec<u8>> {
+) -> Result<Vec<u8>, Error> {
 	let module = wasmi::Module::from_buffer(code)?;
 	let module = instantiate_module(heap_pages, &module)?;
 
 	call_in_wasm_module(ext, &module, method, data)
 }
 
-fn get_mem_instance(module: &ModuleRef) -> Result<MemoryRef> {
+fn get_mem_instance(module: &ModuleRef) -> Result<MemoryRef, Error> {
 	Ok(module
 		.export_by_name("memory")
 		.ok_or_else(|| Error::InvalidMemoryReference)?
@@ -345,7 +348,7 @@ fn get_mem_instance(module: &ModuleRef) -> Result<MemoryRef> {
 
 /// Find the global named `__heap_base` in the given wasm module instance and
 /// tries to get its value.
-fn get_heap_base(module: &ModuleRef) -> Result<u32> {
+fn get_heap_base(module: &ModuleRef) -> Result<u32, Error> {
 	let heap_base_val = module
 		.export_by_name("__heap_base")
 		.ok_or_else(|| Error::HeapBaseNotFoundOrInvalid)?
@@ -360,12 +363,12 @@ fn get_heap_base(module: &ModuleRef) -> Result<u32> {
 }
 
 /// Call a given method in the given wasm-module runtime.
-pub fn call_in_wasm_module(
+fn call_in_wasm_module(
 	ext: &mut dyn Externalities<Blake2Hasher>,
 	module_instance: &ModuleRef,
 	method: &str,
 	data: &[u8],
-) -> Result<Vec<u8>> {
+) -> Result<Vec<u8>, Error> {
 	call_in_wasm_module_with_custom_signature(
 		ext,
 		module_instance,
@@ -388,8 +391,8 @@ pub fn call_in_wasm_module(
 
 /// Call a given method in the given wasm-module runtime.
 fn call_in_wasm_module_with_custom_signature<
-	F: FnOnce(&mut dyn FnMut(&[u8]) -> Result<u32>) -> Result<Vec<RuntimeValue>>,
-	FR: FnOnce(Option<RuntimeValue>, &MemoryRef) -> Result<Option<R>>,
+	F: FnOnce(&mut dyn FnMut(&[u8]) -> Result<u32, Error>) -> Result<Vec<RuntimeValue>, Error>,
+	FR: FnOnce(Option<RuntimeValue>, &MemoryRef) -> Result<Option<R>, Error>,
 	R,
 >(
 	ext: &mut dyn Externalities<Blake2Hasher>,
@@ -397,7 +400,7 @@ fn call_in_wasm_module_with_custom_signature<
 	method: &str,
 	create_parameters: F,
 	filter_result: FR,
-) -> Result<R> {
+) -> Result<R, Error> {
 	// extract a reference to a linear memory, optional reference to a table
 	// and then initialize FunctionExecutor.
 	let memory = get_mem_instance(module_instance)?;
@@ -439,10 +442,10 @@ fn call_in_wasm_module_with_custom_signature<
 }
 
 /// Prepare module instance
-pub fn instantiate_module(
+fn instantiate_module(
 	heap_pages: usize,
 	module: &Module,
-) -> Result<ModuleRef> {
+) -> Result<ModuleRef, Error> {
 	// start module instantiation. Don't run 'start' function yet.
 	let intermediate_instance = ModuleInstance::new(
 		module,
@@ -465,6 +468,209 @@ pub fn instantiate_module(
 	}
 }
 
+/// A state snapshot of an instance taken just after instantiation.
+///
+/// It is used for restoring the state of the module after execution.
+#[derive(Clone)]
+struct StateSnapshot {
+	/// The offset and the content of the memory segments that should be used to restore the snapshot
+	data_segments: Vec<(u32, Vec<u8>)>,
+	/// The list of all global mutable variables of the module in their sequential order.
+	global_mut_values: Vec<RuntimeValue>,
+	heap_pages: u64,
+}
+
+impl StateSnapshot {
+	// Returns `None` if instance is not valid.
+	fn take(
+		module_instance: &ModuleRef,
+		data_segments: Vec<DataSegment>,
+		heap_pages: u64,
+	) -> Option<Self> {
+		let prepared_segments = data_segments
+			.into_iter()
+			.map(|mut segment| {
+				// Just replace contents of the segment since the segments will be discarded later
+				// anyway.
+				let contents = mem::replace(segment.value_mut(), vec![]);
+
+				let init_expr = match segment.offset() {
+					Some(offset) => offset.code(),
+					// Return if the segment is passive
+					None => return None
+				};
+
+				// [op, End]
+				if init_expr.len() != 2 {
+					return None;
+				}
+				let offset = match init_expr[0] {
+					Instruction::I32Const(v) => v as u32,
+					Instruction::GetGlobal(idx) => {
+						let global_val = module_instance.globals().get(idx as usize)?.get();
+						match global_val {
+							RuntimeValue::I32(v) => v as u32,
+							_ => return None,
+						}
+					}
+					_ => return None,
+				};
+
+				Some((offset, contents))
+			})
+			.collect::<Option<Vec<_>>>()?;
+
+		// Collect all values of mutable globals.
+		let global_mut_values = module_instance
+			.globals()
+			.iter()
+			.filter(|g| g.is_mutable())
+			.map(|g| g.get())
+			.collect();
+
+		Some(Self {
+			data_segments: prepared_segments,
+			global_mut_values,
+			heap_pages,
+		})
+	}
+
+	/// Reset the runtime instance to the initial version by restoring
+	/// the preserved memory and globals.
+	///
+	/// Returns `Err` if applying the snapshot is failed.
+	fn apply(&self, instance: &ModuleRef) -> Result<(), WasmError> {
+		let memory = instance
+			.export_by_name("memory")
+			.ok_or(WasmError::ApplySnapshotFailed)?
+			.as_memory()
+			.cloned()
+			.ok_or(WasmError::ApplySnapshotFailed)?;
+
+		// First, erase the memory and copy the data segments into it.
+		memory
+			.erase()
+			.map_err(|_| WasmError::ApplySnapshotFailed)?;
+		for (offset, contents) in &self.data_segments {
+			memory
+				.set(*offset, contents)
+				.map_err(|_| WasmError::ApplySnapshotFailed)?;
+		}
+
+		// Second, restore the values of mutable globals.
+		for (global_ref, global_val) in instance
+			.globals()
+			.iter()
+			.filter(|g| g.is_mutable())
+			.zip(self.global_mut_values.iter())
+			{
+				// the instance should be the same as used for preserving and
+				// we iterate the same way it as we do it for preserving values that means that the
+				// types should be the same and all the values are mutable. So no error is expected/
+				global_ref
+					.set(*global_val)
+					.map_err(|_| WasmError::ApplySnapshotFailed)?;
+			}
+		Ok(())
+	}
+}
+
+/// A runtime along with its version and initial state snapshot.
+#[derive(Clone)]
+pub struct WasmiRuntime {
+	/// A wasm module instance.
+	instance: ModuleRef,
+	/// Runtime version according to `Core_version`.
+	///
+	/// Can be `None` if the runtime doesn't expose this function.
+	version: Option<RuntimeVersion>,
+	/// The snapshot of the instance's state taken just after the instantiation.
+	state_snapshot: StateSnapshot,
+}
+
+impl WasmiRuntime {
+	/// Perform an operation with the clean version of the runtime wasm instance.
+	fn with<R, F>(&self, f: F) -> R
+		where
+			F: FnOnce(&ModuleRef) -> R,
+	{
+		self.state_snapshot.apply(&self.instance).expect(
+			"applying the snapshot can only fail if the passed instance is different
+			from the one that was used for creation of the snapshot;
+			we use the snapshot that is directly associated with the instance;
+			thus the snapshot was created using the instance;
+			qed",
+		);
+		f(&self.instance)
+	}
+}
+
+impl WasmRuntime for WasmiRuntime {
+	fn set_heap_pages(&mut self, heap_pages: u64) -> bool {
+		self.state_snapshot.heap_pages == heap_pages
+	}
+
+	fn call(&mut self, ext: &mut dyn Externalities<Blake2Hasher>, method: &str, data: &[u8])
+			-> Result<Vec<u8>, Error>
+	{
+		self.with(|module| {
+			call_in_wasm_module(ext, module, method, data)
+		})
+	}
+
+	fn version(&self) -> Option<RuntimeVersion> {
+		self.version.clone()
+	}
+}
+
+pub fn create_instance<E: Externalities<Blake2Hasher>>(ext: &mut E, code: &[u8], heap_pages: u64)
+	-> Result<WasmiRuntime, WasmError>
+{
+	let module = Module::from_buffer(&code).map_err(|_| WasmError::InvalidModule)?;
+
+	// Extract the data segments from the wasm code.
+	//
+	// A return of this error actually indicates that there is a problem in logic, since
+	// we just loaded and validated the `module` above.
+	let data_segments = extract_data_segments(&code)?;
+
+	// Instantiate this module.
+	let instance = instantiate_module(heap_pages as usize, &module)
+		.map_err(WasmError::Instantiation)?;
+
+	// Take state snapshot before executing anything.
+	let state_snapshot = StateSnapshot::take(&instance, data_segments, heap_pages)
+		.expect(
+			"`take` returns `Err` if the module is not valid;
+				we already loaded module above, thus the `Module` is proven to be valid at this point;
+				qed
+				",
+		);
+
+	let version = call_in_wasm_module(ext, &instance, "Core_version", &[])
+		.ok()
+		.and_then(|v| RuntimeVersion::decode(&mut v.as_slice()).ok());
+	Ok(WasmiRuntime {
+		instance,
+		version,
+		state_snapshot,
+	})
+}
+
+/// Extract the data segments from the given wasm code.
+///
+/// Returns `Err` if the given wasm code cannot be deserialized.
+fn extract_data_segments(wasm_code: &[u8]) -> Result<Vec<DataSegment>, WasmError> {
+	let raw_module: RawModule = deserialize_buffer(wasm_code)
+		.map_err(|_| WasmError::CantDeserializeWasm)?;
+
+	let segments = raw_module
+		.data_section()
+		.map(|ds| ds.entries())
+		.unwrap_or(&[])
+		.to_vec();
+	Ok(segments)
+}
 
 #[cfg(test)]
 mod tests {
